+++ fchownat.c
@@ -14,9 +14,6 @@
 
 #include "iamroot.h"
 
-extern uid_t next_getegid();
-extern uid_t next_geteuid();
-
 __attribute__((visibility("hidden")))
 int next_fchownat(int dfd, const char *path, uid_t owner, gid_t group,
 		  int atflags)
@@ -40,6 +37,8 @@ int next_fchownat(int dfd, const char *path, uid_t owner, gid_t group,
 
 int fchownat(int dfd, const char *path, uid_t owner, gid_t group, int atflags)
 {
+	const uid_t oldowner = owner;
+	const uid_t oldgroup = group;
 	char buf[PATH_MAX];
 	ssize_t siz;
 	int ret;
@@ -50,8 +49,8 @@ int fchownat(int dfd, const char *path, uid_t owner, gid_t group, int atflags)
 		return -1;
 	}
 
-	owner = next_geteuid();
-	group = next_getegid();
+	owner = __get_uid();
+	group = __get_gid();
 
 	__debug("%s(dfd: %i, path: '%s' -> '%s', owner: %i, group: %i, atflags: 0x%x)\n",
 		__func__, dfd, path, buf, owner, group, atflags);
@@ -59,6 +58,13 @@ int fchownat(int dfd, const char *path, uid_t owner, gid_t group, int atflags)
 	__remove_at_empty_path_if_needed(buf, atflags);
 	ret = next_fchownat(dfd, buf, owner, group, atflags);
 	__ignore_error_and_warn(ret, dfd, path, atflags);
+	/* Force ignoring EPERM error if not chroot'ed */
+	if ((ret == -1) && (errno == EPERM)) {
+		ret = 0;
+		errno = 0;
+	}
+	__set_uid(buf, oldowner, owner);
+	__set_gid(buf, oldgroup, group);
 
 	return ret;
 }