+++ fstatat64.c
@@ -18,9 +18,6 @@
 #include "iamroot.h"
 
 #ifdef _LARGEFILE64_SOURCE
-extern uid_t next_getegid();
-extern uid_t next_geteuid();
-
 __attribute__((visibility("hidden")))
 int next_fstatat64(int dfd, const char *path, struct stat64 *statbuf,
 		   int atflags)
@@ -68,15 +65,17 @@ int fstatat64(int dfd, const char *path, struct stat64 *statbuf, int atflags)
 	if (ret == -1)
 		goto exit;
 
-	uid = next_geteuid();
-	if (statbuf->st_uid == uid)
-		statbuf->st_uid = geteuid();
+	uid = __get_uid(buf);
+	if (uid == (uid_t)-1)
+		statbuf->st_uid = 0;
 
-	gid = next_getegid();
-	if (statbuf->st_gid == gid)
+	gid = __get_gid(buf);
+	if (gid == (gid_t)-1)
 		statbuf->st_gid = 0;
 
 	__st_mode(buf, statbuf);
+	__st_uid(buf, statbuf);
+	__st_gid(buf, statbuf);
 
 exit:
 	return ret;